feat: load totalCount only for filters with parent
JIRA: TNT-1679
