+++ libs/sdk-ui-filters/src/AttributeFilter/hooks/useAttributeFilterControllerData.ts
@@ -1,7 +1,10 @@
 // (C) 2022 GoodData Corporation
 import { useState, useEffect } from "react";
-import { IMultiSelectAttributeFilterHandler } from "../../AttributeFilterHandler/index.js";
-import { isLimitingAttributeFiltersEmpty } from "../utils.js";
+import {
+    IMultiSelectAttributeFilterHandler,
+    isLimitingAttributeFiltersEmpty,
+} from "../../AttributeFilterHandler/index.js";
+
 import { useAttributeFilterHandlerState } from "./useAttributeFilterHandlerState.js";
 import { MAX_SELECTION_SIZE, PARENT_FILTERS_CORRELATION } from "./constants.js";
 import { filterObjRef } from "@gooddata/sdk-model";
