+++ libs/sdk-ui-filters/src/AttributeFilterHandler/index.ts
@@ -7,3 +7,5 @@ export {
     IMultiSelectAttributeFilterHandlerOptions,
     ISingleSelectAttributeFilterHandlerOptions,
 } from "./factory.js";
+
+export { isLimitingAttributeFiltersEmpty } from "./utils.js";
